+++ app/src/main/java/org/ole/planet/myplanet/repository/SubmissionRepository.kt
@@ -1,0 +1,7 @@
+package org.ole.planet.myplanet.repository
+
+import org.ole.planet.myplanet.model.RealmSubmission
+
+interface SubmissionRepository {
+    fun getPendingSurveys(userId: String?): List<RealmSubmission>
+}
