+++ app/src/main/java/org/ole/planet/myplanet/repository/UserRepository.kt
@@ -1,0 +1,11 @@
+package org.ole.planet.myplanet.repository
+
+import io.realm.Realm
+import org.ole.planet.myplanet.model.RealmUserModel
+
+interface UserRepository {
+    suspend fun getUserProfile(): String?
+    suspend fun saveUserData(data: String)
+    fun getRealm(): Realm
+    fun getCurrentUser(): RealmUserModel?
+}
