+++ app/src/main/java/org/ole/planet/myplanet/repository/UserRepositoryImpl.kt
@@ -1,0 +1,31 @@
+package org.ole.planet.myplanet.repository
+
+import android.content.SharedPreferences
+import io.realm.Realm
+import javax.inject.Inject
+import org.ole.planet.myplanet.datamanager.ApiInterface
+import org.ole.planet.myplanet.datamanager.DatabaseService
+import org.ole.planet.myplanet.model.RealmUserModel
+
+class UserRepositoryImpl @Inject constructor(
+    private val databaseService: DatabaseService,
+    private val preferences: SharedPreferences,
+    private val apiInterface: ApiInterface,
+) : UserRepository {
+
+    override suspend fun getUserProfile(): String? {
+        return preferences.getString("user_profile", null)
+    }
+
+    override suspend fun saveUserData(data: String) {
+        preferences.edit().putString("user_profile", data).apply()
+    }
+
+    override fun getRealm(): Realm {
+        return databaseService.realmInstance
+    }
+
+    override fun getCurrentUser(): RealmUserModel? {
+        return databaseService.realmInstance.where(RealmUserModel::class.java).findFirst()
+    }
+}
